Create code Python to detect wax candles
Note: This code assumes that you have captured an image of the candles with the Raspberry Pi camera and saved it as 'candles.jpg' in the same directory as your Python file. Also, you may need to adjust the threshold values in the `cv2.threshold` function to get the best results depending on the lighting conditions of your environment.